+++ R/boxr_delete_restore.R
@@ -36,7 +36,7 @@ box_restore_file <- function(file_id) {
       "https://api.box.com/2.0/file/",
       file_id
     ),
-    httr::config(token = getOption("boxr.token"))
+    get_token()
   )
   
   if (httr::http_status(req)$message == "Success: (201) Created")
@@ -64,7 +64,7 @@ box_restore_folder <- function(dir_id) {
       "https://api.box.com/2.0/folders/",
       dir_id
     ),
-    httr::config(token = getOption("boxr.token"))
+    get_token()
   )
   
   if (httr::http_status(req)$message == "Success: (201) Created")
@@ -91,7 +91,7 @@ boxDeleteFile <- function(file_id) {
       "https://api.box.com/2.0/files/",
       file_id
     ),
-    httr::config(token = getOption("boxr.token"))
+    get_token()
   )
   
   if (httr::http_status(req)$message == "Success: (204) No Content")
@@ -111,7 +111,7 @@ boxDeleteFolder <- function(dir_id) {
       "https://api.box.com/2.0/folders/",
       dir_id, "?recursive=true"
     ),
-    httr::config(token = getOption("boxr.token"))
+    get_token()
   )
   
   if (httr::http_status(req)$message == "Success: (204) No Content")