+++ tests/testthat/test_02_clear_out.R
@@ -5,13 +5,12 @@ context("Clear out")
 boxr:::create_test_dir()
 
 test_that("Local directory is created", {
-  expect_true(grepl("./test_dir", list.dirs(recursive = F)))
+  expect_true(any(grepl("test_dir", list.dirs(here::here()))))
 })
 
 # Make sure the remote directory in the test account is clear
 test_that("Clear out the remote directory", {
-  skip_on_cran()
-  boxr:::skip_on_travis()
+  skip_if_no_token()
   
   options(boxr.verbose = FALSE)
   # Tell boxr to synch the remote home directory with an empty local one