+++ tests/testthat/test_03_upload_download.R
@@ -10,8 +10,7 @@ context("Upload/download/update")
 writeLines("Completely Original File\n", "test_dir/testfile.txt")
 
 test_that("Uploading a new file", {
-  skip_on_cran()
-  skip_on_travis()
+  skip_if_no_token()
   
   # This dir_id doesn't exist
   expect_error(box_ul(1, "test_dir/testfile.txt"))
@@ -25,8 +24,7 @@ test_that("Uploading a new file", {
 })
 
 test_that("Downloading a file", {
-  skip_on_cran()
-  skip_on_travis()
+  skip_if_no_token()
   
   # Overwrite default to FALSE, local copy present
   expect_error(box_dl(test_vars$file_id, "test_dir"))
@@ -43,8 +41,7 @@ test_that("Downloading a file", {
 
 
 test_that("Updating a file", {
-  skip_on_cran()
-  skip_on_travis()
+  skip_if_no_token()
   options(boxr.verbose = FALSE)
   
   # This dir_id doesn't exist